+++ apps/trumbowyg/assets/trumbowyg/js/trumbowyg.config.js
@@ -13,47 +13,33 @@
 // You should have received a copy of the GNU General Public License
 // along with this file.  If not, see <https://www.gnu.org/licenses/>.
 
-var trumbowygConfig = {
+let trumbowygConfig = {
+    /* TODO: get language from branch-data */
     lang: 'en',
-    btnsDef: {
-        image: {
-            dropdown: ['insertImage', 'upload'],
-            ico: 'insertImage'
-        },
-    },
     btns: [
-        ['viewHTML'],
         ['undo', 'redo'], // Only supported in Blink browsers
         ['formatting'],
         ['highlight'],
         ['strong', 'em', 'del'],
         ['superscript', 'subscript'],
-        ['link'],
         ['unorderedList', 'orderedList'],
+        ['link'],
         ['horizontalRule'],
         ['removeformat'],
-        ['fullscreen']
     ],
     semantic: {
         'b': 'strong',
         'i': 'em',
         's': 'del',
         'strike': 'del',
-        'div': 'div', // Editor does nothing on div
+        'div': 'p',
     },
     // Reduce the link overlay to use only 'url' and 'text' fields,
     // omitting 'title' and 'target'
     minimalLinks: true,
     removeformatPasted: true,
     svgPath: '/static/trumbowyg/dist/ui/icons.svg',
-    plugins: {
-        /* TODO: implement me */
-        upload: {
-            serverPath: '/upload/',
-            fileFieldName: 'image'
-        }
-    },
     // Allow to set link target attribute value to what you want,
     // even if the 'minimalLinks' option is set to true.
     defaultLinkTarget: '_blank',
-}
+};